requirements.txt
absl-py==1.2.0
anyio==3.6.2
APScheduler==3.9.1.post1
astroid==2.12.10
beautifulsoup4==4.11.1
black==22.8.0
bs4==0.0.1
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
dnspython==2.3.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.49.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
numpy==1.23.3
oauthlib==3.2.1
pandas==1.5.2
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.15.3
pymongo==4.3.3
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.3.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
